sysprep the system
we no longer need to hard-code the mac addresses because they
are now automatically generated at the next boot.

we still need to reset the vmk0 mac address to the qemu/machine
mac address to be able to get the VM IP address from qemu/libvirt.
this as the side-effect of restarting the vmk0/management
interface on the first boot and temporary having the initial vmk0
mac address appearing in the network and receiving an IP address
from DHCP.